modify count_one

+++ bit_utils.c
@@ -4,7 +4,7 @@ int neg_value(int a) {
 	return ~a + 1;
 }
 
-int count_one(unsigned int a) {
+int count_one(unsigned long a) {
 	int n = 0;
 	while(a) {
 		if(a & 0x01) n++;
