+++ bit_utils.h
@@ -5,7 +5,7 @@
 int neg_value(int a);
 
 /* count_one counts the number of 1 bits in the binary representation of a */
-int count_one(unsigned int a);
+int count_one(unsigned long a);
 
 #endif
